Move macro-related modules to Frontend
