+++ hs-bindgen/src-internal/HsBindgen/TraceMsg.hs
@@ -30,11 +30,11 @@ import GHC.Generics (Generic)
 
 import Clang.HighLevel.Types (Diagnostic (..))
 import HsBindgen.BindingSpec (BindingSpecMsg (..))
-import HsBindgen.C.Reparse.Infra (ReparseError (..))
-import HsBindgen.C.Tc.Macro (TcMacroError (..))
 import HsBindgen.Clang (ClangMsg (..))
 import HsBindgen.Frontend (FrontendMsg (..))
 import HsBindgen.Frontend.Analysis.DeclIndex (DeclIndexError (..))
+import HsBindgen.Frontend.Macro.Reparse.Infra (ReparseError (..))
+import HsBindgen.Frontend.Macro.Tc (TcMacroError (..))
 import HsBindgen.Frontend.Pass.HandleMacros.IsPass (HandleMacrosMsg (..))
 import HsBindgen.Frontend.Pass.HandleTypedefs.IsPass (HandleTypedefsMsg (..))
 import HsBindgen.Frontend.Pass.MangleNames.IsPass (MangleNamesMsg (..))
